+++ src/lte/model/lte-rlc-sequence-number.h
@@ -78,7 +78,7 @@ public:
   SequenceNumber10 operator++ (int)
   {
     SequenceNumber10 retval (m_value);
-    m_value = (m_value + 1) % 1024;
+    m_value = ((uint32_t)m_value + 1) % 1024;
     return retval;
   }
